Move EventBus outside masks, remove a re-export index file
