+++ packages/react-data-grid/src/ReactDataGrid.tsx
@@ -10,7 +10,7 @@ import React, {
 
 import Grid from './Grid';
 import { getColumnMetrics } from './ColumnMetrics';
-import { EventBus } from './masks';
+import EventBus from './EventBus';
 import { CellNavigationMode, EventTypes, UpdateActions, HeaderRowType, DEFINE_SORT } from './common/enums';
 import {
   AddFilterEvent,
