+++ packages/react-data-grid/src/RowGroup.tsx
@@ -2,7 +2,7 @@
 import React, { forwardRef } from 'react';
 import { EventTypes } from './common/enums';
 import { CellMetaData, RowRendererProps, CalculatedColumn } from './common/types';
-import EventBus from './masks/EventBus';
+import EventBus from './EventBus';
 
 interface Props<R> {
   height: number;
