--- packages/react-data-grid/src/masks/index.ts
@@ -1,2 +1,0 @@
-export { default as EventBus } from './EventBus';
-export { default as InteractionMasks } from './InteractionMasks';
